Fixed EnPassant move. Fixed pawn 2 space kill bug. Implemented promotionTo request and response (left test with front update).
